+++ CHANGELOG.md
@@ -304,4 +304,5 @@
 * IGUI panels can be added with a builder function instead of prefab
 * Config panels now inform user about pending changes
 * Added Coroutine methods to AsyncUtils
+* Added RunOnMainThread methods to AsyncUtils [Experimental]
 * QdmsMessageComponent can now translate and rebroadcast as Unity messages [Experimental]